<Custom-Flight-Controller> Updated STM32 HAL Library
Changes:
1. Updated STM32 HAL Library
2. Updated InterruptMgr. Now interrupt manager is at high level and it uses direct API call to disseminate.
3. Added SBUS Driver. NOT TESTED.

+++ FlightController/FlightController/Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_cryp_ex.h
@@ -2,13 +2,11 @@
   ******************************************************************************
   * @file    stm32f4xx_hal_cryp_ex.h
   * @author  MCD Application Team
-  * @version V1.6.0
-  * @date    04-November-2016
   * @brief   Header file of CRYP HAL Extension module.
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; COPYRIGHT(c) 2016 STMicroelectronics</center></h2>
+  * <h2><center>&copy; COPYRIGHT(c) 2017 STMicroelectronics</center></h2>
   *
   * Redistribution and use in source and binary forms, with or without modification,
   * are permitted provided that the following conditions are met:
@@ -67,10 +65,10 @@
 /** @defgroup CRYPEx_Exported_Constants_Group1 CRYP AlgoModeDirection
   * @{
   */ 
-#define CRYP_CR_ALGOMODE_AES_GCM_ENCRYPT   ((uint32_t)0x00080000U)
-#define CRYP_CR_ALGOMODE_AES_GCM_DECRYPT   ((uint32_t)0x00080004U)
-#define CRYP_CR_ALGOMODE_AES_CCM_ENCRYPT   ((uint32_t)0x00080008U)
-#define CRYP_CR_ALGOMODE_AES_CCM_DECRYPT   ((uint32_t)0x0008000CU)
+#define CRYP_CR_ALGOMODE_AES_GCM_ENCRYPT   0x00080000U
+#define CRYP_CR_ALGOMODE_AES_GCM_DECRYPT   0x00080004U
+#define CRYP_CR_ALGOMODE_AES_CCM_ENCRYPT   0x00080008U
+#define CRYP_CR_ALGOMODE_AES_CCM_DECRYPT   0x0008000CU
 
 /**
   * @}
@@ -80,7 +78,7 @@
   * @brief    The phases are relevant only to AES-GCM and AES-CCM
   * @{
   */ 
-#define CRYP_PHASE_INIT           ((uint32_t)0x00000000U)
+#define CRYP_PHASE_INIT           0x00000000U
 #define CRYP_PHASE_HEADER         CRYP_CR_GCM_CCMPH_0
 #define CRYP_PHASE_PAYLOAD        CRYP_CR_GCM_CCMPH_1
 #define CRYP_PHASE_FINAL          CRYP_CR_GCM_CCMPH
@@ -100,8 +98,8 @@
 /**
   * @brief  Set the phase: Init, header, payload, final. 
   *         This is relevant only for GCM and CCM modes.
-  * @param  __HANDLE__: specifies the CRYP handle.
-  * @param  __PHASE__: The phase.
+  * @param  __HANDLE__ specifies the CRYP handle.
+  * @param  __PHASE__ The phase.
   * @retval None
   */
 #define __HAL_CRYP_SET_PHASE(__HANDLE__, __PHASE__)  do{(__HANDLE__)->Instance->CR &= (uint32_t)(~CRYP_CR_GCM_CCMPH);\
